+++ src/Stroytorg.App/libs/stroytorg-assets/src/lib/styles/component-styles/stroytorg-timepicker.scss
@@ -2,7 +2,7 @@
   display: none;
   position: absolute;
   top: 2.5em;
-  left: -3em;
+  left: 0;
   width: 16.5em;
   min-width: 8em;
   max-width: 80dvw;
