+++ src/Stroytorg.App/libs/stroytorg-components/src/lib/components/stroytorg-date/components/stroytorg-datepicker/stroytorg-datepicker.component.html
@@ -53,14 +53,14 @@
       <stroytorg-select
         formControlName="month"
         [items]="viewFacade.monthOptions"
-        [size]="viewFacade.selectSize.SMALL"
+        [inputSize]="viewFacade.selectSize.SMALL"
         [clearable]="false"
       ></stroytorg-select>
       <stroytorg-select
         formControlName="year"
         [items]="viewFacade.yearOptions"
         [bindLabel]="'value'"
-        [size]="viewFacade.selectSize.XSMALL"
+        [inputSize]="viewFacade.selectSize.XSMALL"
         [clearable]="false"
       ></stroytorg-select>
     </div>
