+++ src/Stroytorg.App/libs/stroytorg-components/src/lib/components/stroytorg-date/stroytorg-date.component.html
@@ -2,7 +2,7 @@
 <div
   class="d-flex justify-content-center align-items-center align-items-center gap-1"
 >
-  <div class="stroytorg-date-container {{ size }}">
+  <div class="stroytorg-date-container {{ inputSize }}">
     <input
       class="stroytorg-date"
       type="date"
