+++ src/Stroytorg.App/libs/stroytorg-components/src/lib/components/stroytorg-date/stroytorg-date.models.ts
@@ -1,10 +1,1 @@
-export enum DateSize {
-  XSMALL = 'x-small-width',
-  SMALL = 'small-width',
-  DEFAULT = 'default-width',
-  LARGE = 'large-width',
-  XLARGE = 'x-large-width',
-  FULL = 'full-width',
-};
-
 export const TODAY = new Date();
