+++ src/Stroytorg.App/libs/stroytorg-components/src/lib/components/stroytorg-select/index.ts
@@ -1,2 +1,1 @@
 export * from './stroytorg-select.component';
-export * from './stroytorg-select.models';
