+++ src/Stroytorg.App/libs/stroytorg-components/src/lib/components/stroytorg-select/stroytorg-select.component.html
@@ -1,6 +1,6 @@
 @if(formControl){
 <div class="d-flex justify-content-center align-items-center align-items-center gap-1">
-  <div class="stroytorg-select-container {{ size }}">
+  <div class="stroytorg-select-container {{ inputSize }}">
     <ng-select
       floatingHint
       class="stroytorg-select"
