+++ src/Stroytorg.App/libs/stroytorg-components/src/lib/components/stroytorg-text-input/stroytorg-text-input.component.html
@@ -2,7 +2,7 @@
 <div
   class="d-flex justify-content-center align-items-center align-items-center gap-1"
 >
-  <div class="stroytorg-text-input-container {{ size }}">
+  <div class="stroytorg-text-input-container {{ inputSize }}">
     @if(type === 'singleline'){
     <input
       floatingHint
