+++ src/Stroytorg.App/libs/stroytorg-components/src/lib/components/stroytorg-text-input/stroytorg-text-input.component.ts
@@ -2,7 +2,7 @@ import { ChangeDetectionStrategy, Component, EventEmitter, Input, Optional, Outp
 import { CommonModule } from '@angular/common';
 import { ControlValueAccessor, NgControl, ReactiveFormsModule } from '@angular/forms';
 import { StroytorgBaseFormInputComponent, StroytorgBaseInputControls } from '../stroytorg-base-form';
-import { InputSize, InputType } from './stroytorg-text-input.models';
+import { InputType } from './stroytorg-text-input.models';
 import { ErrorPipe, FloatingHintDirective } from '@stroytorg/shared';
 
 @Component({
@@ -27,14 +27,7 @@ export class StroytorgTextInputComponent
     super(ngControl);
   }
 
-  @Input()
-  size = InputSize.DEFAULT as string;
-
   @Output() valueChange = new EventEmitter<string>();
 
-  id = (Math.random() + 1).toString(36).substring(7);
-
   @Input() type = InputType.SINGLELINE as string;
-
-  InputSize = this.size;
 }
