+++ src/Stroytorg.App/libs/stroytorg-components/src/lib/components/stroytorg-text-input/stroytorg-text-input.models.ts
@@ -1,10 +1,3 @@
-export enum InputSize {
-  DEFAULT = 'default-width',
-  LARGE = 'large-width',
-  XLARGE = 'x-large-width',
-  FULL = 'full-width',
-}
-
 export enum InputType {
   MULTILINE = 'multiline',
   SINGLELINE = 'singleline',
