+++ src/Stroytorg.App/libs/stroytorg-components/src/lib/components/stroytorg-time/stroytorg-time.component.html
@@ -2,7 +2,7 @@
 <div
   class="d-flex justify-content-center align-items-center align-items-center gap-1"
 >
-  <div class="stroytorg-time-container {{ size }}">
+  <div class="stroytorg-time-container {{ inputSize }}">
     <input
       class="stroytorg-time"
       type="time"
