--- src/Stroytorg.App/libs/stroytorg-components/src/lib/components/stroytorg-time/stroytorg-time.models.ts
@@ -1,10 +1,0 @@
-export enum TimeSize {
-  XSMALL = 'x-small-width',
-  SMALL = 'small-width',
-  DEFAULT = 'default-width',
-  LARGE = 'large-width',
-  XLARGE = 'x-large-width',
-  FULL = 'full-width',
-}
-
-// export const NOW: Time = new Date().getTime();
